For Sakai 2.0.1/CTools 2.2:
From Sakai 1.5 to Saka 2.0, there are a bunch of changes in tool ids. 
chef.* i->sakai.*
sakai.noti.prefs -> sakai.preferences
sakai.presence -> sakai.online
sakai.siteinfogeneric -> sakai.siteinfo
sakai.sitesetupgeneric -> sakai.sitesetup
sakai.threadeddiscussion -> sakai.discussion

The problem was that if the batch import in Admin Archive tool was done against the old exported file (site.xml), a new site was created with site id as its title and empty tool set.

This fix is in Archive service that when merging site tool information, it update the tool id inside the DOM if it has the old format/naming, and write it back to DOM. 



git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@103 66ffb92e-73f9-0310-93c1-f5514f145a0a