For Sakai 2.0.1 and CTools 2.2
Fix to SAK-841: Email Archive Tool's content should be imported when using Site Archive Tool.

The archived email message is marked from "postmaster". Since "postmaster" is not a user in the target site, no email message is imported.

The fix ignores the user check when the message is marked from "postmaster".

git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@134 66ffb92e-73f9-0310-93c1-f5514f145a0a