SAK-1216 SAK-1216
This corrects the display of "TypeException" messages in the "list" view of the resources tool for the creator of resources in sites that are imported if the creator is not known on the server to which the site is imported.
It also corrects a similar problem (also covered by SAK-1216) in the edit and properties views of the resources tool for those sites.




git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@264 66ffb92e-73f9-0310-93c1-f5514f145a0a