In reference to: SAK-1244
Not very happy about this. Changed the scrolling attribute of the iframe that displays messages to "auto"

Normal messages will still display as usual. These edge case messages will get the scrollbar. Ugly, but at least they will be able to get at the content - which is likely to be long pasted in urls.





git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@307 66ffb92e-73f9-0310-93c1-f5514f145a0a