Related to SAK-1445
Removed unused files related to tool's titles and layout, to reduce confusion.  The portal does the layout of the title and main frame of tools.  A individual tool can have its own layout internally in its main frame, but the portal knows nothing about that.

Also get it compiling better in Eclipse.


git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@348 66ffb92e-73f9-0310-93c1-f5514f145a0a