SAK-1572 http://bugs.sakaiproject.org/jira/browse/SAK-1572
Comments of submitter: "The [>] button in the list navigation isn't on the same line as [<] and the pull-down menu. The problem persists regardless of window size. I don't see the same problem with other list navigation, so I'm guessing it's a bug.

On Safari, by the way, it's the Previous button [<] that is on a separate line, so it appears at least semi-browser dependent."

Fix - removed white-space:nowrap of all forms that make up this navigator

git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@891 66ffb92e-73f9-0310-93c1-f5514f145a0a