Fixes SAK-1507: Enabling DEBUG logging breaks Sakai
Check for a null SessionManager and handle it gracefully.  The SessionManager may be null sometimes when the server is starting up.

This bug is now a known issue in 2.0.0 and 2.0.1 and should be added to the list of known issues.  However, since it only affects developers (NOT production), it should be noted that its not a huge issue.

git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@1444 66ffb92e-73f9-0310-93c1-f5514f145a0a