SAK-1450 SAK-1483 SAK-2726 - outgoing email header improvements
SMTP Port control
------------------------

You can set the port as well as the SMTP server to use for outgoing email, by configuring the EmailService either in components.xml or the sakai.properties.
To set the port, set the smtpPort property in the bean definition, or use this in sakai.properties:

smtpPort@org.sakaiproject.service.framework.email.EmailService=NNNN

where "NNNN" is the port number; the default is 25.


Message-Id Preservation
----------------------------------

The Message-Id header from incoming email will be preserved in the email sent out to site participants.  So will the Date header.


Date Headers
-------------------

If otherwise not set, outgoing messages will have a Date header set to the current time.


Return Path
----------------

Messages will have a return path specified, by setting the MAIL FROM value in the SMTP envelope.  The value can be configured by setting the smtpFrom property
in the bean definition, or this in sakai.properties:

smtpPort@org.sakaiproject.service.framework.email.EmailService=MAILADDR

where "MAILADDR" is the email address to send errors, bounces, vacation messages, etc.

If this is not set, it defaults to "postmaster@" your Sakai server.  Usually, this is just quietly eaten in the incoming mail processing.

if you are *very* careful, you can establish a Site with an Email Archive tool with the "postmaster" alias (or name the site "postmaster")
and messages will accumulate there.  Make sure that there are NO users who have access to the site, or else the bounced mail will be
reflected to these users (and you are in deep trouble if any of these bounce...).  Also make sure that the email channel is set to accept
mail from ANY user, or else you will get another kind of bouncing going on.  There's real potential for a mail loop

if you do this, so proceed with great care.


Promotion of SMPT settings to Java System Properties
-------------------------------------------------------------------------

The following system properties are now set as soon as the EmailService starts up, based on the configuration and default values of the service:

mail.smtp.host
mail.smtp.port
mail.smtp.from

These are done *before* that other email processor (i.e. James) has a chance to screw things up.  It should no longer be necessary to set these on
the java command line with -D options.




git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@3714 66ffb92e-73f9-0310-93c1-f5514f145a0a