SAK-2025
Before 2.1, if a user had multiple accounts in Sakai, all with the same email address, they may not be able to send emails to a site that they are members of and otherwise would have permissions to.  This is because Sakai's UserDirectoryService (and UserDirectoryProvider) has a findUserByEmail() method, which is used by the email processing to identify which user is sending in the mail, and which finds only one user.  If there are many users with the same email address, it finds one of them.

The problem arises when it finds a user account that does not have permission, even if there's another that does.  This method is not consistent with Sakai's policy about email addresses in user records; they are not unique.

Sakai 2.1 fixes this by changing the UserDirectoryService method to findUsersByEmail(), returning a Collection of all the users with this email.  Email processing uses this to identify all the possible  users who might have sent the an email, and then allows the email to come in if any of these potential users have permission.

There are two consequences of this change.  One is a database "change" (sort of), enforcing new case rules about email addresses to improve lookup efficiency.  The other is a new API (sort of) for the UserDirectoryProvider, to allow it to support multiple user records per email.

* * *

Email rules call for case sensitive user names and case in-sensitive domain parts of the email address.  Sakai now enforces this, buy manipulating any email addresses going into the User record (preserve the case of the user id part, lower the case of the domain), and any string used to search for emails.  This now allows us to do precise searching, using database indexes, avoiding a performance problem as user databases grow in size.

The database schema is unchanged, but if you have records from before Sakai 2.1, you might have some non-lower case domain names in user email addresses.  You will NOT see this from inside Sakai; you have to look from your back-end database connection.

Any user record that has a non-lower case domain name in the email can be fixed by editing and saving from Sakai, either from the user's account tool or the admin user editor.  To process all your records, look for a special update button on the Admin user editor.

If old records with non-lowercase domain names remain, those users will not be able to send email into Sakai - their email address will never be found to be any user's email address.

* * *

UserDirectoryProviders that have only a single user record for any email address need not change.  If you want to support non-unique email addresses; multiple user records with the same email address, you need to do the following:

Add one more interface to your UDP component's "implements" clause:

	public class SampleUserDirectoryProvider implements UserDirectoryProvider, UsersShareEmailUDP

For the UsersShareEmailUDP interface:

public interface UsersShareEmailUDP
{
	/**
	 * Find all user objects which have this email address.
	 *
	 * @param email
	 *        The email address string.
	 * @param factory
	 *        Use this factory's newUser() method to create all the UserEdit objects you populate and return in the return collection.
	 * @return Collection (UserEdit) of user objects that have this email address, or an empty Collection if there are none.
	 */
	Collection findUsersByEmail(String email, UserFactory factory);
}

you need to add one more method to your implementation.  See the sample user directory provider for an example.  This method returns a collection, either empty if no user records have the email address, or full of UserEdit objects, which are created by calling the factory parameter, and filled in by your component, for all the user records that match the email address.

Note: the email search string is already "normalized" to have case-preserved name and case-lowered domain.

If you implement this new method, you can use an empty implementation for findUserByEmail() (or throw a not implemented exception), as this really should not be called.



git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@3865 66ffb92e-73f9-0310-93c1-f5514f145a0a