Fix to SAK-1051: Customized page order gets reset if tools are added using worksite setup
It was a problem when there was no toolOrder.xml file in sakai folder.
Then the rules for tool ordering through Site Info tool "Edit Tools" process are:
1) Don't change the ordering of any existing site page;
2) For any newly added tools, they are to be put towards the end of tool list; Alpha ordered.
3) Home and Site Info tools are exceptions. When Home is added, it gets moved to the top of list; When Site Info tool is added, it is moved to the end.
4) However, if Home or Site Info is of existing tools, don't change its order.

When a new site is created from Worksite Setup, the tools are alpha-orded. Home tool, if chosen, is at top. Site Info, if chosen, is at bottom.

Remember, the tool order we are discussing here is under the senario of:
1) there is no toolOrder.xml specified;Otherwise, the order from that file is used.
2) we are not considering the Help tool, which is mandated, always appears at the bottom of tool list in a site, and not selectable from WSetup or SiteInfo tool.


git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@4121 66ffb92e-73f9-0310-93c1-f5514f145a0a