SAK-3145 - enable tool stealthing via sakai.properties or tool manager component configuration.
The org.sakaiproject.api.kernel.tool.ActiveToolManager can be configured with the "stealthTools" property,
a list of comma-separated tool ids (no spaces) for all tools that are to be stealthed.

This can be overridden in sakai.properties with:

# stealth tools.  A stealth tool is a tool that is running in Sakai, but is not available to be added to a site in Worksite Setup.
# A stealth tool can be added to a site using the Admin site tool.
# List any tools to be stealthed, using their tool ids, in a comma-separated string (no spaces).
# In this example, the rwiki tool, su tool, roster tool, and the assignment tool w/o grading are stealthed:
stealthTools@org.sakaiproject.api.kernel.tool.ActiveToolManager=sakai.rwiki,sakai.su,sakai.site.roster,sakai.assignment

Stealthed tools will not be listed by the tool manager when findTools() is called with a category - this has the effect of ignoring
these tools' categories as defined in their tool.xml (without the need to edit those tool.xml files).  This results in the tools
not showing up in Worksite Setup's tool lists, but showing up in the Admin Site Editor.



git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@4512 66ffb92e-73f9-0310-93c1-f5514f145a0a