Reverse merged: SAK-2330 http://bugs.sakaiproject.org/jira/browse/SAK-2330 Display exceptions in stead of a blank page
I think the problem with this code is it's in the wrong place.  Coincidentally, I recently was studying how to better do exception handling in a request, and where to inform the end user.  I didn't need that work for the problem I was working on, but it lead me to a better understanding of the issue.

This is a HARD issue, btw.

I believe the problem is a portal problem, not a kernel / active tool manager one.  Fixing it there, and then smoothing all the path between any possible error and the portal code is the way to fix this.


git-svn-id: https://source.sakaiproject.org/svn/trunk/sakai@5788 66ffb92e-73f9-0310-93c1-f5514f145a0a