From: jimeng Date: August 5, 2006 10:39:17 AM EDT Subject: Re: please create branch for availability features
Lance,

It includes new features that will be used in production at Boston
University, Whitman College, UVISA and possibly other institutions
beginning in three weeks or so, once they complete QA.  The new
features (ability to hide resources from most users and ability to
schedule release and retraction of resources) will be part of 2.3.

By the time the QA cycle is done, I expect it will be based on 2.2.1,
but for now, it is based on 2.2.0 because that is already a known
quantity.

Clay is managing the QA with help from Francette and others.

Jim


git-svn-id: https://source.sakaiproject.org/svn/sakai/branches/sakai_pre-2.3_availability_features@13414 66ffb92e-73f9-0310-93c1-f5514f145a0a